Place

+++ src/main/java/com/capstonebe/capstonebe/place/controller/PlaceController.java
@@ -21,13 +21,12 @@ public class PlaceController {
     private final UserService userService;
     private final JwtUtil jwtUtil;
 
-    // todo : 관리자 인증
     // 장소 등록
     @PostMapping("/register")
     public ResponseEntity<?> registerPlace(@RequestBody @Valid PlaceRegisterRequest request,
                                            @CookieValue(value = "jwt", required = false) String token) {
 
-        //validateAdmin(token);
+        validateAdmin(token);
 
         return ResponseEntity.ok(placeService.registerPlace(request));
     }
@@ -37,7 +36,7 @@ public class PlaceController {
     public ResponseEntity<?> editPlace(@RequestBody @Valid PlaceEditRequest request,
                                        @CookieValue(value = "jwt", required = false) String token) {
 
-        //validateAdmin(token);
+        validateAdmin(token);
 
         return ResponseEntity.ok(placeService.editPlace(request));
     }
@@ -47,7 +46,7 @@ public class PlaceController {
     public ResponseEntity<?> deletePlace(@PathVariable Long id,
                                          @CookieValue(value = "jwt", required = false) String token) {
 
-        //validateAdmin(token);
+        validateAdmin(token);
 
         placeService.deletePlace(id);
